# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API сервиса подачи заявок на 3D-печать МИЭМ НИУ ВШЭ.\\nАутентификация — OTP через email + JWT (Bearer token).",
        "title": "MIEM 3D Lab Print API",
        "contact": {},
        "version": "1.0.0"
    },
    "host": "localhost:3000",
    "basePath": "/",
    "paths": {
...
        "/api/auth/logout": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Выйти (отозвать refresh token)",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.LogoutRequest"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/refresh": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Обновить токены",
                "parameters": [
                    {
                        "description": "Refresh token",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.RefreshRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.RefreshResponse"
                        }
                    },
                    "401": {
                        "description": "Невалидный/просроченный/отозванный token",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/request-otp": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Запросить OTP-код",
                "parameters": [
                    {
                        "description": "Email адрес",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.RequestOTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.RequestOTPResponse"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "429": {
                        "description": "Превышен лимит запросов",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    }
                }
            }
        },
        "/api/auth/verify-otp": {
            "post": {
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "auth"
                ],
                "summary": "Подтвердить OTP-код",
                "parameters": [
                    {
                        "description": "Email + OTP-код",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.VerifyOTPRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_dto.VerifyOTPResponse"
                        }
                    },
                    "400": {
                        "description": "Неверный или истёкший код",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "404": {
                        "description": "OTP не найден",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "423": {
                        "description": "Слишком много попыток",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "miem_3d_lab_print_backend_src_internal_apierr.APIError": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string",
                    "example": "INTERNAL_ERROR"
                },
                "details": {
                    "type": "object"
                },
                "message": {
                    "type": "string",
                    "example": "Внутренняя ошибка сервера"
                }
            }
        },
        "miem_3d_lab_print_backend_src_internal_apierr.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "$ref": "#/definitions/miem_3d_lab_print_backend_src_internal_apierr.APIError"
                }
            }
        },
...
        "miem_3d_lab_print_backend_src_internal_dto.LogoutRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
...
        "miem_3d_lab_print_backend_src_internal_dto.RefreshRequest": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "miem_3d_lab_print_backend_src_internal_dto.RefreshResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "expires_in": {
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                }
            }
        },
        "miem_3d_lab_print_backend_src_internal_dto.RequestOTPRequest": {
            "type": "object",
            "required": [
                "email"
            ],
            "properties": {
                "email": {
                    "type": "string"
                }
            }
        },
        "miem_3d_lab_print_backend_src_internal_dto.RequestOTPResponse": {
            "type": "object",
            "properties": {
                "expires_in": {
                    "type": "integer"
                },
                "message": {
                    "type": "string"
                }
            }
        },
...
        "miem_3d_lab_print_backend_src_internal_dto.VerifyOTPRequest": {
            "type": "object",
            "required": [
                "code",
                "email"
            ],
            "properties": {
                "code": {
                    "type": "string"
                },
                "email": {
                    "type": "string"
                }
            }
        },
        "miem_3d_lab_print_backend_src_internal_dto.VerifyOTPResponse": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string"
                },
                "consent_required": {
                    "type": "boolean"
                },
                "expires_in": {
                    "type": "integer"
                },
                "refresh_token": {
                    "type": "string"
                },
                "role": {
                    "type": "string"
                },
                "token_type": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Введите \"Bearer \u003caccess_token\u003e\"",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}